classwork/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
isOdd==0.1.2
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
progress==1.6
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
python-telegram-bot==20.0